+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 - Disallowed `Screen` instances in `App.SCREENS` and `App.MODES`
+- Fixed `App.MODES` being the same for all instances -- per-instance modes now exist under `app.modes`
+
 
 ### Added